fix(contracts): remove dead models.py, consolidate enum validation (QW-09)
- Delete src/elspeth/core/landscape/models.py (393 LOC duplicate)
- Move _validate_enum helper to contracts/audit.py
- Add __post_init__ validators to Run, Node, Edge for Tier-1 crash behavior
- Update test import from models to contracts

The models.py file was a complete duplicate of contracts/audit.py.
All dataclass contracts now have a single source of truth with
proper enum validation that crashes on invalid types per Data Manifesto.
